Code re-struct and cleanup for auth and reg device
