+++ MockRegSBI/app/src/main/AndroidManifest.xml
@@ -47,8 +47,7 @@
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </activity>
-        <activity android:name="nprime.reg.mocksbi.device.RCaptureActivity" />
-        <activity android:name="nprime.reg.mocksbi.device.AuthCaptureActivity" />
+        <activity android:name="nprime.reg.mocksbi.device.CaptureActivity" />
 
         <provider
             android:name="androidx.core.content.FileProvider"
